+++ src/data/items.ts
@@ -1,0 +1,11 @@
+export const BASIC_ITEMS: Record<string, any> = {
+  'bread_crust': { id: 'bread_crust', name: "Stale Bread Crust", type: 'consumable', rarity: 'common', description: "Hard as a rock and speckled with mold, but hunger makes it a feast. It's the standard ration at the orphanage, barely enough to keep a child alive.", value: 1, weight: 0.1, stats: { health: 2, stamina: 5 } },
+  'coin': { id: 'coin', name: "Gold Coin", type: 'misc', rarity: 'common', description: "The currency of the realm. Cold, hard, and necessary. It bears the faded profile of a long-dead emperor, a reminder of the empire's waning influence.", value: 1, weight: 0.01 },
+  'wooden_sword': { id: 'wooden_sword', name: "Wooden Sword", type: 'weapon', rarity: 'common', description: "A child's toy, splintered and worn from countless imaginary battles. It offers little protection, but holding it brings a fleeting sense of courage.", value: 2, weight: 1, stats: { willpower: 5 } },
+  'blue_mountain_flower': { id: 'blue_mountain_flower', name: "Blue Mountain Flower", type: 'consumable', rarity: 'common', description: "A common flower with minor restorative properties. Its petals are a vibrant, unnatural blue, hinting at the latent magic that permeates the land.", value: 2, weight: 0.1, stats: { health: 5, purity: 1 } },
+  'glowing_mushroom': { id: 'glowing_mushroom', name: "Glowing Mushroom", type: 'consumable', rarity: 'uncommon', description: "A strange, bioluminescent mushroom found in damp caves and cellars. It hums with faint, unsettling magic that makes your skin prickle when you hold it.", value: 5, weight: 0.2, stats: { health: -5, lust: 10, corruption: 2 } },
+  'rusty_iron_dagger': { id: 'rusty_iron_dagger', name: "Rusty Iron Dagger", type: 'weapon', rarity: 'common', description: "A discarded, rusted blade found in the muck. The edge is dull, chipped, and stained with questionable brown spots. It's barely sharp enough to cut cheese, but gripping its worn, sweat-stained leather hilt gives you a slight sense of security in these dark alleys. It smells faintly of old blood, rust, and desperation.", value: 5, weight: 2, stats: { willpower: 10 } },
+  'threadbare_tunic': { id: 'threadbare_tunic', name: "Threadbare Tunic", type: 'clothing', rarity: 'common', description: "A simple, coarse wool tunic. It's thin, itchy, and offers almost no protection from the cold or anything else. It smells faintly of sweat and the desperate, cramped conditions of the orphanage. Wearing it immediately marks you as someone of low social standing.", value: 1, weight: 0.5 },
+  'strange_amulet': { id: 'strange_amulet', name: "Strange Amulet", type: 'misc', rarity: 'rare', description: "A heavy, cold metal amulet depicting an eye surrounded by tentacles. It whispers to you in the dark when you try to sleep.", value: 50, weight: 0.5, stats: { willpower: -5, corruption: 5 } },
+  'healing_poultice': { id: 'healing_poultice', name: "Healing Poultice", type: 'consumable', rarity: 'uncommon', description: "A foul-smelling paste made of crushed herbs and mud. It burns when applied, but it closes wounds quickly.", value: 10, weight: 0.5, stats: { health: 20, pain: -10 } }
+};
